#18 refactor: 컨트롤러 구조 변경
- TestRedisController를 RedisHealthController로 이름 변경
- Redis 관련 패키지 경로 변경
- 새로운 MongoHealthController 추가
- HealthCheckController의 패키지 경로 변경
