+++ src/main/kotlin/com/millo/ollim/health/controller/HealthCheckController.kt
@@ -1,4 +1,4 @@
-package com.millo.ollim
+package com.millo.ollim.health.controller
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.tags.Tag
